# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HR Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    put:
      operationId: updateEmployee
      tags: [employees]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /employees/{id}/counters:
    put:
      operationId: updateEmployeeCounters
      tags: [employees]
      summary: Update an employee's running counters
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCountersBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"

  # ── OFFICES ────────────────────────────────────────────────────────────────
...
  /monthly-configs/{year}/{month}:
    get:
      operationId: getMonthlyConfig
      tags: [monthlyConfig]
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: integer
        - name: month
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyConfig"
    put:
      operationId: upsertMonthlyConfig
      tags: [monthlyConfig]
      summary: Create or update monthly configuration
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: integer
        - name: month
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertMonthlyConfigBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyConfig"

  # ── PUBLIC HOLIDAYS ────────────────────────────────────────────────────────
...
  /settings/mail:
    get:
      operationId: getMailSettings
      tags: [mailSettings]
      summary: Get SMTP mail settings (admin only)
      responses:
        "200":
          description: Mail settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MailSettings"
    put:
      operationId: updateMailSettings
      tags: [mailSettings]
      summary: Update SMTP mail settings (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MailSettings"
      responses:
        "200":
          description: Updated mail settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MailSettings"
...
components:
  schemas:
...
    MailSettings:
      type: object
      properties:
        host:
          type: string
          nullable: true
        port:
          type: integer
          nullable: true
        secure:
          type: boolean
          nullable: true
        username:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        fromAddress:
          type: string
          nullable: true
        enabled:
          type: boolean

    # DEPARTMENT
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          type: string
          description: "lu | be | de | fr | other"
        contractPercent:
          type: number
          description: "e.g. 100, 80, 50"
        weeklyContractHours:
          type: number
          description: "Contracted hours per week"
        homeworkEligible:
          type: boolean
        coworkEligible:
          type: boolean
        allowedShiftCodes:
          type: array
          items:
            type: string
          description: "Which codes this employee accepts"
        permanenceGroup:
          type: integer
          nullable: true
          description: "1 or 2, null if not in a permanence group"
        permanenceLevel:
          type: integer
          nullable: true
          description: "1 (primary) or 2 (escalation)"
        isSpoc:
          type: boolean
        spocRotates:
          type: boolean
        isManagement:
          type: boolean
        prmCounter:
          type: number
          description: "Running PRM balance in hours"
        holidayHoursRemaining:
          type: number
          description: "Remaining holiday hours"
        overtimeHours:
          type: number
        homeworkDaysUsedThisYear:
          type: integer
          description: "TT days used — relevant for cross-border workers (max 35/year)"
        dayCodePreferences:
          type: array
          nullable: true
          items:
            type: object
            properties:
              day:
                type: integer
                description: "0=Mon, 1=Tue, 2=Wed, 3=Thu, 4=Fri"
              code:
                type: string
            required: [day, code]
          description: "Favourite shift codes per weekday used by the auto-planner"
        prefersHeightAdjustableDesk:
          type: boolean
          nullable: true
          description: "Whether the employee prefers a height-adjustable desk"
        departmentId:
          type: integer
          nullable: true
          description: "ID of the department this employee belongs to"
        preferredOfficeId:
          type: integer
          nullable: true
          description: "ID of the office this employee prefers to work from (used by the auto-planner)"
        onsiteWeekRatio:
          type: number
          nullable: true
          description: "Custom onsite week ratio for the auto-planner (e.g. 0.5 = 50%). Null uses the default formula (floor(weeks/2))."
        displayOrder:
          type: integer
          description: "Manual sort order for the planning view. Lower numbers appear first within a department group."
        role:
          type: string
          nullable: true
          description: "Employee organisational role (e.g. analyst, manager, team lead)"
        email:
          type: string
          nullable: true
          description: "Employee email address (used for demand notifications)"
        approverAdminId:
          type: integer
          nullable: true
          description: "ID of the admin user who approves demands for this employee"
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - country
        - contractPercent
        - weeklyContractHours
        - homeworkEligible
        - coworkEligible
        - allowedShiftCodes
        - isSpoc
        - spocRotates
        - isManagement
        - prmCounter
        - holidayHoursRemaining
        - overtimeHours
        - homeworkDaysUsedThisYear
        - displayOrder
        - createdAt
        - updatedAt
    CreateEmployeeBody:
      type: object
      properties:
        name:
          type: string
        country:
          type: string
        contractPercent:
          type: number
        weeklyContractHours:
          type: number
        homeworkEligible:
          type: boolean
        coworkEligible:
          type: boolean
        allowedShiftCodes:
          type: array
          items:
            type: string
        permanenceGroup:
          type: integer
          nullable: true
        permanenceLevel:
          type: integer
          nullable: true
        isSpoc:
          type: boolean
        spocRotates:
          type: boolean
        isManagement:
          type: boolean
        prmCounter:
          type: number
        holidayHoursRemaining:
          type: number
        overtimeHours:
          type: number
        homeworkDaysUsedThisYear:
          type: integer
        departmentId:
          type: integer
          nullable: true
        preferredOfficeId:
          type: integer
          nullable: true
          description: "Preferred office for onsite planning"
        onsiteWeekRatio:
          type: number
          nullable: true
          description: "Custom onsite week ratio (0.0–1.0). Null = use default formula."
        notes:
          type: string
          nullable: true
      required:
        - name
        - country
        - contractPercent
        - weeklyContractHours
        - homeworkEligible
        - coworkEligible
        - allowedShiftCodes
        - isSpoc
        - spocRotates
        - isManagement
    UpdateEmployeeBody:
      type: object
      properties:
        name:
          type: string
        country:
          type: string
        contractPercent:
          type: number
        weeklyContractHours:
          type: number
        homeworkEligible:
          type: boolean
        coworkEligible:
          type: boolean
        allowedShiftCodes:
          type: array
          items:
            type: string
        permanenceGroup:
          type: integer
          nullable: true
        permanenceLevel:
          type: integer
          nullable: true
        isSpoc:
          type: boolean
        spocRotates:
          type: boolean
        isManagement:
          type: boolean
        dayCodePreferences:
          type: array
          nullable: true
          items:
            type: object
            properties:
              day:
                type: integer
                description: "0=Mon, 1=Tue, 2=Wed, 3=Thu, 4=Fri"
              code:
                type: string
            required: [day, code]
        prefersHeightAdjustableDesk:
          type: boolean
          nullable: true
        departmentId:
          type: integer
          nullable: true
        preferredOfficeId:
          type: integer
          nullable: true
          description: "Preferred office for onsite planning"
        onsiteWeekRatio:
          type: number
          nullable: true
          description: "Custom onsite week ratio (0.0–1.0). Null = use default formula."
        displayOrder:
          type: integer
          description: "Manual sort order within the planning view group."
        role:
          type: string
          nullable: true
          description: "Employee organisational role (e.g. analyst, manager, team lead)"
        email:
          type: string
          nullable: true
          description: "Employee email address"
        approverAdminId:
          type: integer
          nullable: true
          description: "ID of the admin user who approves demands for this employee"
        notes:
          type: string
          nullable: true
    UpdateCountersBody:
      type: object
      properties:
        prmCounter:
          type: number
        holidayHoursRemaining:
          type: number
        overtimeHours:
          type: number
        homeworkDaysUsedThisYear:
          type: integer

    # OFFICE
...
    MonthlyConfig:
      type: object
      properties:
        id:
          type: integer
        year:
          type: integer
        month:
          type: integer
        contractualHours:
          type: number
          description: "Total hours to plan this month (from PRMS doc)"
        jlDays:
          type: integer
          description: "Number of CCT-FHL JL days available this month (distributed by the planner)"
        notes:
          type: string
          nullable: true
      required: [id, year, month, contractualHours, jlDays]
    UpsertMonthlyConfigBody:
      type: object
      properties:
        contractualHours:
          type: number
        jlDays:
          type: integer
          description: "Number of CCT-FHL JL days available this month"
        notes:
          type: string
          nullable: true
      required: [contractualHours, jlDays]

    # PUBLIC HOLIDAYS
...